Functioning model wrapper
Still kind of a mess but it works.

One thing that came up while wrapping this is that there isn't any real
batching in the model. Batching is implemented by calling the forward
method in a loop using one of the wrapper functions. It looks like that
would require significant refactoring so it's probably better to handle
that in a pipeline component for this.